Add files via upload
This folder contains all the necessary files to reproduce the figures and tables presented in the manuscript titled "A Bayesian mark interaction model for analysis of tumor pathology images", as well as its associated supplement. Source the R code titled with a specific figure or table name/index, and you will visualize the data or the analysis result exactly same with the one in the manuscript or the supplement.